[msd-arm-mali] Improve robustness of power on/off waits.
Improve the robustness of the power on/off wait routines by having them
check the status of the operation in 1ms increments, instead of simply
waiting a full second to timeout. It has been observed that, with some
low probability, the expected wake ups do not occur even though the
status changes to the desired state. These hangs are frequent enough to
fail automated quality metrics.

Bug: 115768
Bug: b/242564647
Change-Id: I996d513c0929c98d427105639deffc62e3c70800