[mdns] query for SRV/TXT/A/AAAA when not sent proactively
Normally, when an instance is discovered via a PTR query, the
responder sends SRV, TXT and A/AAAA resources along with the PTR
resource. When this doesn't happen, the requestor should ask for
the missing resources explicitly. The mdns service was not asking for
missing records. This CL addresses that.

If the requestor receives PTR resource with no SRV, it
requests an SRV resource and waits for it to arrive before
notifying the client that the instance is discovered.

If the requestor receives PTR and SRV resources with no A/AAAA, it
requests those records and waits for the resources to arrive before
notifying the client that the instance is discovered.

If the requestor receives PTR and SRC resources with no TXT, it
requests those records and notifies the client of the discovered
instance with an empty list of text strings. If and when the TXT
resource is received, the client is notified that the discovered
instance has changed, and the text string are included in the
instance.

Fixed: b/260003827

Test: new unit tests

Change-Id: I5e251fe219e1e548f06e2b3448856ac2cc94e09f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/764443
Commit-Queue: Dale Sather <dalesat@google.com>
Reviewed-by: Sundarajan Srinivasan <sundarsrini@google.com>
(cherry picked from commit c6d26d38c98ebdb90a6665bf614b3e3cc60aae6b)