[camera] Avoid calling GetDeviceInfo twice
GetDeviceInfo to camera controller was being called twice using two
separate connections. Creating a 2nd connection means that the 2nd
connection is vulnerable to the 1st connection not necessarily
completely disconnected. Also, there is no reason to call the same
FIDL call twice when expecting the results to be the same. The call
is a query for static information. One call is enough.

Bug: fxb/111252
Test: Run 500 cycles to ensure zero SetUI camera probe failures.
Change-Id: If3405dbe524b507144bcee61ef8705489445f66a