[kernel][syscalls] Use timeslice extension in msg_put_handles.
Most places that take the handle table lock for write already use
timeslice extensions to avoid getting preempted while holding the lock
for a short time. This case in the channel_write path showed up in
traces recently and has the same lock thrashing signature as the other
cases addressed previously. Apply the same fix to address the lock
thrash in channel_write.

Bug: b/236891952
Bug: b/242564647
Bug: b/234776870
Change-Id: Ibd10ac48a5c0b62df1a328f73636b765104ba41b