[board][nelson] Add a separate GPIO device for bank C GPIOs
Add a GPIO device that only exposes GPIOs from bank C. This bank uses
separate registers from all other GPIOs, so no synchronization is
required between this and the main GPIO instance.

The exception to this is the ot-radio interrupt GPIO -- the GPIO C
device will not be able to provide interrupts for GPIOs it exposes, so
this GPIO must be exposed by the main device. The interrupt GPIO is only
configured once during init, making synchronization in the ot-radio
driver trivial.

By being a platform device rather than a protocol implementation device,
this GPIO driver instance can be colocated with the SPI and ot-radio
drivers in order to reduce the rate of errors in the CPU coex test.

Bug: 117797
Test: device-enumeration-test passed on Nelson
Test: Thread CPU coex test
Change-Id: If4db225a3d9c6e196d3a9b061af162fbdcc656b1