[recovery] Explicitly switch to direct viewmode.
When carnelian is in ViewMode::Auto, there is a chance the display
controller has not been loaded when carnelian checks.
Since recovery is intended to always run in direct display mode, it can
always use that config. ViewMode::Direct already uses a vfswatcher to
wait for a valid display controller handle.

Test: CQ tests, boots 10/10 with new config on nelson
Bug: b/280111541

Change-Id: I18674559c32442e07db91c8a8f22481c0bd219a1