[brcmfmac] Always initialize banjo structs before use
Previously in certain code paths we were converting uninitialized Banjo
structs to FIDL before making a FIDL request to wlanif. In particular,
this meant that bools may have some underlying value other than 0 or 1,
which would cause it to fail FIDL validation and the server end to
WlanFullmacImplIfc to close. This resulted in a persistent error in
which wlancfg could no longer communicate with the brcmfmac driver.

Test: Manual test. First we forced the brcmfmac driver to send an
uninitialzied struct and verified that the wlanif channel would close.
Then we exercised the same code path but with the struct initialized and
verified that we did not see any issues.

Bug: fxb/135685
Change-Id: Ifc04af64ce5d8566483ae6dbbedbf2ba81e416f5