[superproject] Roll third_party/Vulkan-Tools [logs] Depend explicitly on syslog library
Previously, this was indirectly depending on the syslog C++ library.
This used to work when syslog was non-hermetic due to not being
exported to the SDK. When we export to the SDK, it will need to be
made hermetic which breaks symbol resolution.

Original-Bug: b/300157652
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/third_party/Vulkan-Tools/+/949072
Original-Revision: 9f0f40b6429aed1a8371142bfc7e31adf3d07c30
GitOrigin-RevId: 1fa3f88fd73cf9b1f4622c77eae7aeff484943c7
Roller-URL: https://ci.chromium.org/b/8749291025308687521
CQ-Do-Not-Cancel-Tryjobs: true
Change-Id: I2c62afe8f9dde5600aaab22397f683399c65be21