[superproject] Roll fuchsia [starnix][ltp] Temporarily skip eventfd tests
These tests will pass as soon as we add CONFIG_EVENTFD to the kconfig
for the LTP tests. However, the kconfig file is in another repo, so we
need to skip these tests while we make the transition.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/976332
Original-Revision: b110dc2a684db3f2de47ceab5f7b2a59db652747
GitOrigin-RevId: 50908e5e32fb0a8f9168dd1f84cfe4e6a2cac816
Roller-URL: https://ci.chromium.org/b/8749258664034426977
CQ-Do-Not-Cancel-Tryjobs: true
Change-Id: I68df2672d460b4f5973d30ba1615c4aa89286e0f