[superproject] Roll fuchsia [bazel_sdk] Don't set soname in fuchsia_cc_driver
If a soname is set, the loader looks at the soname to determine if the
shared library is already loaded, and if so returns the old copy. This
prevents us from loading a new version of a driver if an old version
already exists.

Original-Bug: b/326445083
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/995435
Original-Revision: 48130ad3a3bad23f6525596c0a18d83fc4547b20
GitOrigin-RevId: 7db4f18854ceedef86bff7754e34134eb98e7d26
Roller-URL: https://ci.chromium.org/b/8749234323631552961
CQ-Do-Not-Cancel-Tryjobs: true
Change-Id: I054bac02feaaad765b32e8dde9c32bc8b7c2dd53