[ffx] fix target list --format a
Make the output of `ffx target list --format a` exactly compatible with
`ffx target get-ssh-address`, which is going away. This means that it
will always include the port, rather than doing so only when it is not
22 (which is the format for the default `ffx target list` output). This
seems reasonable since the user is explicitly asking for an address, so
it should probably explicitly provide the address.

Change-Id: I6a6a6964bf4811cfdc3fbaffda04186936b8e68e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1338905
Reviewed-by: Andrew Davies <awdavies@google.com>
Commit-Queue: Steven Grady <slgrady@google.com>