[ffx] Implement usb_driver command
This provides the actual implementation of the USB driver.

The library to let a client talk to the driver is in a forthcoming CL.
It's hard to unit-test this code without that so we don't yet, but tests
are coming.

Fixed: 429029313
Change-Id: I8063d4a0212cb064973d38ab52c563b314de4075
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1314432
Fuchsia-Auto-Submit: Casey Dahlin <sadmac@google.com>
Reviewed-by: Ian McKellar <ianloic@google.com>
Commit-Queue: Casey Dahlin <sadmac@google.com>
API-Review: Ian McKellar <ianloic@google.com>
Reviewed-by: Andrew Davies <awdavies@google.com>