[kernel][vm] Add VmMapping::DestroyLockedObject helper
To simplify merging VmMappings, this introduces a helper that performs
the destruction of a VmMapping, but with the option to not perform the
unmap. This allows for the page tables to remain, whilst the VmMapping
is still removed from the VMO and VMAR trees.

This new helper requires the VmObject lock to be held, which is
required for most of the destruction path anyway. This does add extra
unecessary work into the VmObject lock, but only trivial amounts and
the additional code simplifications and reuse are worth it.

Change-Id: I4d70c5eb245e6a42ed360e30094a16a1b777c542
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1323347
Commit-Queue: Adrian Danis <adanis@google.com>
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>