[kernel][vm] Always create new mappings when unmapping
Previously VmMapping::Unmap would attempt to either shorten or move an
existing mapping to fulfill the unmap request. This is efficient, but
also complex and rigid.

To simplify this, Unmap is changed to always create new mappings for the
remaining regions and destroy the existing one, which allows for a much
simpler implementation.

This is less efficient as it requires always performing new
allocations and tree insertions, but is worth the simplification and
will allow for the base_ and size_ of a mapping to eventually become
constant.

Change-Id: Ieaa14eac60822d47365ffe1711ff2e376b9f5324
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1323348
Commit-Queue: Adrian Danis <adanis@google.com>
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>