[kernel][vm] Merging mappings creates a new mapping
Instead of modifying an existing VmMapping when merging, a new mapping
is created and the old ones are destroyed. This is less efficient as it
requires a new allocation, but it is a rare enough operation that the
simplification is worth it.

This change also makes it easier to make the base and size of a VmMapping
constant in the future.

Change-Id: I6f2b7d8995dbb80276c0390799d5f3f437387d73
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1323349
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>
Commit-Queue: Adrian Danis <adanis@google.com>