[moonflower][audio] Introduce config fuchsia.starnix.ExpectMcuReady.
The motivation for this config flag is to signal that the MCU is
expected to become operational.

The audio driver needs this information to decide whether to wait for
the MCU to become operational and provide audio resources (e.g.,
microphone). Without starnix (e.g., on minimal.sorrel), the audio driver
should not expect to use such audio resources.

Bug: 377393100

Change-Id: I76630d909798d12eb7a973e51ff83fa89a2be9ed
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1335643
Reviewed-by: Andres Oportus <andresoportus@google.com>
Commit-Queue: HanBin Yoon <hanbinyoon@google.com>
Reviewed-by: Aaron Wood <aaronwood@google.com>