[libc][dl][test] Remove decls in test base class
This removes the few function declarations in DlTestsBase. Their purpose
was to provide some documentation of the API that child test classes
should support, but in practice it's not that useful, and its easy to
forget to update this class with new supported functions.

Bug: 356458230
Change-Id: I0ba4c2b4b6d6fa9eb8d9102f801df9d53e3b09ee
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1338906
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Fuchsia-Auto-Submit: Caslyn Tonelli <caslyn@google.com>
Commit-Queue: Caslyn Tonelli <caslyn@google.com>