[ffx][trace] Include & fix lib/trace tests
It looks like we were inadvertently not building the lib/trace tests and
as a result they bitrotted a little. I discovered this when I went to
add more tests and found that the test target didn't even build.

The one bug that I found was that the category validation regular
expression incorrectly allowed double-quotes.

Bug: 433532694
Change-Id: Ia83be35780afcbe86c11a60b9fccc988e0d7da08
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1337847
Reviewed-by: Gwen Mittertreiner <gmtr@google.com>
Commit-Queue: Ian McKellar <ianloic@google.com>
Reviewed-by: Casey Dahlin <sadmac@google.com>