[paver][moonflower] support paving to any partition
Adds support to the moonflower paver to write images to arbitrary
partitions. This is useful now so that we can add more images to the OTA
package later and the paver will know what to do with them.

Without this change, the behavior would be to silently ignore any
unknown files. In some cases this is what we want, but for moonflower,
this would have caused problems if we switch over to a different boot
image format where all the partitions need to be written together or
else boot will fail. Silently ignoring images only works if the image is
standalone and doesn't depend on any other part of the OTA, which is not
the case here.

Bug: b/435725164
Change-Id: I3dc16202b7e172d0303eebe73fd1de7f6f700a62
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1338665
Reviewed-by: Yecheng Zhao <zyecheng@google.com>
Fuchsia-Auto-Submit: David Pursell <dpursell@google.com>
Commit-Queue: David Pursell <dpursell@google.com>