[kernel][vm] ZeroPagesLocked assert not hidden
`ZeroPagesLocked` is currently only called from VMOs so there should
be no cases of it being called on a hidden VmCowPages. I also think that
it would be broken if called on a hidden VmCowPages so I added this
assert to fire in case that happens for some reason.

Change-Id: Ic36691a102ddb62d8cbbc8f2dfbd6e29f8dbb89f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1337564
Commit-Queue: Sage Barreda <sagebarreda@google.com>
Reviewed-by: Ethan Williams <ethanws@google.com>