[kernel][crashlog] Clarify dependency on register dump formatting
We never formalized a scheme for register dump formatting and versioning
and have no dependents on the current version in need of one. While the
original inclusion of "(v1.0)" was certainly well-intentioned, it is
currently premature and has been a source of confusion (foremost on the
part of this author) around its actual relevance, and is a stumbling
block to expanding the dump. Accordingly, we remove the mention.

What is something we have implicit dependency on though is the
"REGISTERS" logging before the dump in our crashlog tests. To clarify
that and make it slightly more robust we deduplicate the responsibility
for printing that by moving it from arch_render_crashlog_registers() to
crashlog_to_string() with a scare quoute mentioning the dependency.

NO_IFTTT=static analysis contradicts official documentation around
specifying multiple ThenChange files.

Change-Id: Ia3714accb40b34f0adaedca4120f49e1b3f371ea
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1340931
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Francois Rousseau <frousseau@google.com>
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>