[unwinder] CfiParser recognizes 32bit transition frames
...when coming from Starnix. Specifically, this means that we examine
both PC and LR (in aarch64 terms) and guess that if either of them are
within the 32 bit address space, that we need to construct a 32 bit
register set for them to pass along to the next frame we attempt to
unwind.

Tests are difficult to write at the moment due to the existing testing
strategy not allowing us to collect backtraces of 32 bit binaries. A
follow up will add limited testing coverage to zxdb_e2e_tests, and full
testing support will be covered in https://fxbug.dev/430036274.

Bug: 395606674
Bug: 416040479
Bug: 430036274
Change-Id: I26909fcc92764f7ac334073171bab7c38487d1dd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1272198
Commit-Queue: Kevin Lindkvist <lindkvist@google.com>
Reviewed-by: Adam Barth <abarth@google.com>