[kernel][crashlog] Extend reg dumps for arm64 and x86
It was recently convenient to have tpidr_el1 available to an arm64
crashlog register dump, so we add that here. Then, for parity, we add
tpidr_el0 and tpidrr0_el0 for arm64 as well, and the for further
thread ABI pointer parity with x86 we had fsbase, gsbase, and
'swapgs gsbase'.

Since there were some attempts at lining up the ":" in the dump lines
previously, we follow through on that fully.

Change-Id: I9a2d8c7eb948afdf8f1bbcbff3794677d77f906c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1340213
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Reviewed-by: Travis Geiselbrecht <travisg@google.com>
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>