[cm_graph] DependencyNode owns its data
Currently `DependencyNode<T>` has a `Copy` bound on `T`. In practice `T`
is a `&str`, e.g. a reference to a string in a component declaration for
example. Change this to require `Clone` but not `Copy`, and update the
implementation to pass `&T` instead of `T` to avoid performance
regression.

This allows cm_graph to move from DependencyNode<&str> to
DependencyNode<FlyStr>. This is needed in https://fxrev.dev/1335611
which stores a list of DependencyNode after the referent - a dynamic
offer - is gone.

 # Performance

We use `FlyStr` to minimize the overhead of storing owned data, and even
then it's transient since the `DependencyNode`s are dropped once the
graph computation is complete.

Bug: 426886212

Change-Id: I27bbc013bcece8bea23f35f47d5806d88b42b23c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1333843
Reviewed-by: Olivia Nordquist <nolivia@google.com>
Commit-Queue: Gary Bressler <geb@google.com>