[netstack3] Introduce DataAvailable mechanism
Introduce DataNotifier and DataWatcher types that together implement a
multiple-producer, single-consumer notification mechanism. The notifier
can be cloned and shared, but there is only one watcher. The watcher
must initialize a wait operation in order to be notified and will only
receive notifications that are sent after the operation has been
initialized.

This is meant to be used for sockets in Core to be able to signal the
availability of data to a wake group in Bindings.

Fixed: 434262472
Test: netstack3-bin_test
Change-Id: I4b82b8b1504df488e0a28bfe5155a58a7538af07
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1340204
Commit-Queue: Peter Johnston <peterjohnston@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Fuchsia-Auto-Submit: Peter Johnston <peterjohnston@google.com>