[ui] starnix touch device registers KEY_SLEEP
The touch attribute includes BTN_MISC as a catchall, in addition to
KEY_SLEEP, which the device is programmed to send in certain cases.

The keyboard attribute also includes KEY_VOLUMEDOWN for completeness, as
the keyboard device is also programmed to send this event in certain
cases.

Bug: 406225311
Test: manual test with tqr/1009550
Change-Id: I5faed169b8e4f82422833e97dafb1854bea784b3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1340264
Reviewed-by: Jianpeng Chao <chaopeng@google.com>
Commit-Queue: Caroline Liu <carolineliu@google.com>