[bazel][idk] cquery to generate prebuild info for Bazel atoms
Add a cquery file that generates prebuild info from IDK atoms defined in
Bazel that is equivalent to that generated by the GN build. The only
difference is that labels are in the Bazel format and end in "_idk"
instead of "_sdk".

Modify `additional_prebuild_info` to allow its values to be output in
their original types.

Bug: 417304469
Test: fx bazel cquery build/bazel/bazel_idk/tests:test-source-set_idk --output=starlark --starlark:file=build/bazel/bazel_idk/idk_prebuild_info.cquery
Test: `fx bazel cquery //sdk/lib/magma_common:magma_common_idk --output=starlark --starlark:file=build/bazel/bazel_idk/idk_prebuild_info.cquery` matches the GN prebuild info except the label format.
Change-Id: I55833a39fde5e040efb2db67f30326df559dc54c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1336910
Reviewed-by: Jay Zhuang <jayzhuang@google.com>
Commit-Queue: David Dorwin <ddorwin@google.com>
Reviewed-by: David Turner <digit@google.com>