[bazel][idk] Define target for Bazel atoms in the IDK
Define a "fuchsia_idk_atoms_produced_by_main_platform_build" in Bazel
and add source sets that have been migrated to it.

This target is not yet used and all atoms still come from GN, but it
provides a non-test target to, for example, use with cquery and provides
a basis for thinking about how this will work.

Bug: 417304469
Test: fx bazel build //sdk:fuchsia_idk_atoms_produced_by_main_platform_build
Change-Id: Ie6966f02df79d81f4a57eb71a10584519ecd6128
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1338911
Reviewed-by: David Turner <digit@google.com>
Reviewed-by: Jay Zhuang <jayzhuang@google.com>
Commit-Queue: David Dorwin <ddorwin@google.com>