[sdmmc][mmc] Check packed command status
Per the eMMC specification, the device is allowed to send random data in
place of a packed read that failed. Thus, the only way to to tell if a
packed command failed is to check the status register in EXT_CSD.

For now any packed command errors are ignored to maintain the old
behavior. This check may be reverted if it does not show up in the
b/427293779 crash logs.

Bug: 427293779
Test: Ran biotime on Sorrel to excercise packed commands
Change-Id: I56c716810414d3c5562e19be0e9321af68c6e85f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1337724
Reviewed-by: HanBin Yoon <hanbinyoon@google.com>
Commit-Queue: Braden Kell <bradenkell@google.com>