[kernel][vm] Make VmMapping base and size const
The base and size of a VmAddressRegion are already effectively const.
With the base and size of a VmMapping no longer needing to be changed,
its base and size can also be made const officially in the common base
class.

The rest of this commit is then cleaning up all the previous accessor
variants.

Change-Id: Ie805be890bd0d8296564f9aaf9622e8712402ace
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1323350
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>
Commit-Queue: Adrian Danis <adanis@google.com>