[driver][metadata] Accept svc dir as parameter
Let the various metadata retrieval functions within the
//sdk/lib/driver/metadata/cpp library accept a service directory as well
as a shared pointer of an incoming directory. This will make it easier
to write tests for drivers that use the metadata library because the
driver test framework does not provide the outgoing directory of the
driver being tested but does provide its service directory.

Change-Id: Ifb0e03afcb42507a54ad3d4d1dff19f5ea05dab6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1341844
Reviewed-by: Novin Changizi <novinc@google.com>
Reviewed-by: Sarah Chan <spqchan@google.com>
Commit-Queue: Nick Eaton <nickeaton@google.com>