[audio][adr] Allow retry on create ring buffer failure
When the driver returns error for CreateRingBuffer, the device was
considered to be in error state and was removed. This assumption is
incorrect as the CreateRingBuffer failure could be due to temporary
issues like underlying hardware is not ready or there was a temporary
memory shortage. Updated code to log error on CreateBufferFailure
unless there is a framework/fatal error, in which case the device
is marked as in error.

Test: fx test audio_services_device
Bug: 435504328
Change-Id: Ie931690bd3e25d61ccd43e34a640c2934757edd6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1335114
API-Review: Mukesh Agrawal <quiche@google.com>
Reviewed-by: Martin Puryear <mpuryear@google.com>
Commit-Queue: Puneetha Ramachandra <puneetha@google.com>
Fuchsia-Auto-Submit: Puneetha Ramachandra <puneetha@google.com>
Reviewed-by: Mukesh Agrawal <quiche@google.com>