[starnix][mm] Remove userfault registration from Mapping
We don't need to hold a back pointer from every Mapping object to a
userfault object. Instead, we can use two bits in the Mapping object to
annotate whether the mapping is registered with a userfaultfd and then
keep a list of all the active userfaultfd objects in the
MemoryManagerState (typically there are zero or very few of them).

Bug: b/435494969
Change-Id: Ie4b13bd1a9be0a5405bd61c66aac5448afe42042
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1340619
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>
Fuchsia-Auto-Submit: Adam Barth <abarth@google.com>
Commit-Queue: Adam Barth <abarth@google.com>