[ffx][fuchsia-controller] Log even without the daemon.
This makes it so that anything using ssh will log to the ssh log file
iff the log level is set to DEBUG or higher verbosity.

This change is to aid in debugging of failures in ffx.

This does some minor cleanup of some existing functions:

1. It removes the newline character from the line buffer function so
   that it adheres to the same API as tokio::Lines<>.
2. write_ssh_log is no longer async (as it's not necessary).
3. The ssh command itself appends `-vv` instead of the caller modifying
   its arguments when the logging level is set to DEBUG or higher
   verbosity.

This change also adds the logging ID to the ssh output so that it is
clearer which command/process/library wrote to it (though this could be
improved to show the source and the instance ID to remove the guesswork,
most users debugging ssh will start with the ffx/fuchsia-controller logs
and then proceed with the ID they've found into the ssh file to see
where those logs are).

There is potential future work here to provide some kind of ssh log
separation to make it clearer where the source of the logs came from,
but that will require some discussion.

Fixed: 405204725
Change-Id: I16fde6ee44f1648354865e253975aab86a6a636c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1342028
Commit-Queue: Andrew Davies <awdavies@google.com>
Fuchsia-Auto-Submit: Andrew Davies <awdavies@google.com>
Reviewed-by: Steven Grady <slgrady@google.com>