[power] Add a disable parameter to PowerSamplerConfig
The current implementation of create_power_sampler() decides whether to
actually collect power samples or not based on whether the
MEASUREPOWER_PATH environment variable is set, since up until now, that
environment variable has only been set on hosts in the swarming power
pool. However, in the future we will soon start setting the
MEASUREPOWER_PATH environment variable on hosts outside of the power
pool. To aid this transition, this CL adds a new `disable` parameter for
PowerSamplerConfig, which allows the test to disable power collection
even if the MEASUREPOWER_PATH is set.

Bug: 398295446
Change-Id: I0bc50dd87bc21644fa78ccf138f7cd32cc57952d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1341984
Commit-Queue: Michael Brase <mbrase@google.com>
Reviewed-by: Chris Masone <cmasone@google.com>