[build] Write the hash of RBE-downloaded output files to an xattr
This saves a few seconds on builds by allowing reclient to read
the xattr to get the hash of the file, instead of rehashing the
file on every subsequent build.

Before, an xattr was only set in the skip-downloads case, but
this always writes xattrs containing the file digest and size
when RBE is enabled.

Change-Id: Idca0a64d3ade2cc817347177859aa398fabdd630
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1338917
Reviewed-by: David Fang <fangism@google.com>
Commit-Queue: Aaron Wood <aaronwood@google.com>