[audio][test] Tolerate mono I2S
The I2S frame format is, by definition, stereo. At least one
driver defines a DaiFormatSet with only I2S, but then states
that it supports Mono, causing AudioDriverTests to fail.

To unblock deeper testing with AudioDriverTests on this
driver, temporarily loosen this particular check, with a
constexpr flag kEnforceI2sIsStereoOnly (and accompanying
TODO) that will be removed once the driver issue is fixed.

This CL also defines a test case that can be quickly
commented-in, to display the DaiFormatSets that the driver
exposes. This differs from the DaiFormatSets displayed by
`ffx audio device info`, because those are filtered and
translated through the ADR.

Bug: 437737013

Change-Id: I338ad3649f7f1e04ffebdc86f0a4845eb9d615d9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1342418
Fuchsia-Auto-Submit: Martin Puryear <mpuryear@google.com>
Reviewed-by: HanBin Yoon <hanbinyoon@google.com>
Commit-Queue: Martin Puryear <mpuryear@google.com>