[starnix] Add track_stub for SETSIG and update test expectations.
The fcntl F_SETSIG flag is not implemented, and there is a newly added
test (FcntlTest.RegularFileOAsync) which is blocking a roll. This test
is failing for the same F_SETSIG call, so I've marked it as an expected
failure and added a track_stub for the missing flag.

Bug: 437951946
Test: CQ / Infra
Change-Id: Iefdc6872d6f28c1ccd4df8af9dfae52949772465
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1342968
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>
Fuchsia-Auto-Submit: Aaron Knobloch <aknobloch@google.com>
Commit-Queue: Aaron Knobloch <aknobloch@google.com>