[bringup] Increase capacity for display coordinator.
Previously, we merged all display coordinator disptachers into
the display coordinator's default driver dispatcher, which increased
the workload of the dispatcher and caused the dispatcher to be
throttled.

This change increases the capacity for the display coordinator
to account for the extra workload.

Bug: 437390782
Change-Id: Iaf58073646e2a92dea60cb9ebc9a2d991bfff640
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1342831
Commit-Queue: Yilong Li <liyl@google.com>
Fuchsia-Auto-Submit: Yilong Li <liyl@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>