[honeydew][wlan] Add query_interfaces method to Wlan affordance
Adds method to query all interfaces from the system at once. This can be
used to obtain all the interfaces that the system has and be able to
know what is available for testing without requesting each one
individually.

This is logic that was migrated from antlion so that it's available to
other test environments.

Change-Id: Id763f5f0481ba44ac793e8368326cd1c216ac5bb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1337664
Reviewed-by: Marc Khouri <mnck@google.com>
Commit-Queue: Charles Celerier <chcl@google.com>
Reviewed-by: Patrick Lu <patricklu@google.com>
Fuchsia-Auto-Submit: Charles Celerier <chcl@google.com>