[test-pilot] update env var processing
This CL changes the way test-pilot handles environment variables. Prior
to this CL, test-pilot would accept a specification of which environment
variables to process. The "--env" option would process all conforming
environment variables, and "--env=foo,bar" would process just
FUCHSIA_FOO and FUCHSIA_BAR. A conforming environment variable is one
that starts with "FUCHSIA_" and which appears in the test config schema.

The scheme implemented in this CL gets rid of the "env" option and uses
a substitution scheme. Of course, on the command line, the shell in
question can substitute values from environment variables already. When
processing JSON files, test-pilot now checks to see whether a test
parameter is being assigned a special object of the form
{ "from_env": "<name>" } or { "try_from_env": "<name>" }. Such objects
are replaced with the value of the environment variable <name> with no
mangling and no regard to the schema. The "from_env" variant fails if
the environment variable is undefined while the "try_from_env" variant
simply skips the assignment.

This new approach has several advantages including allowing the use
of environment variables with no particular prefix (e.g. "FUCHSIA_"),
and ability to assign environment variable values to any parameter (not
just one based on the environment variable name).

Test: updated and changed unit tests

Bug: 327640651
Change-Id: I6799b2b86e7cd955a74ffd3062caeaf8521e5ade
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1340212
Reviewed-by: Chase Latta <chaselatta@google.com>
Reviewed-by: Christopher Johnson <crjohns@google.com>
Reviewed-by: Claire Gonyeo <cgonyeo@google.com>
Commit-Queue: Dale Sather <dalesat@google.com>