[ffx][flash]: unzip product bundle
Product Bundles for certain products need to be able to compare the sha
of the contents. This product bundle code assumes that the product
bundle is a directory, not a zip file. Therefore for ffx to flash a
product bundle for those products, we need to extract them to disk
first.

Bug: b/417799966
Fixed: b/417799966
Change-Id: I53ea139d72828dc9f0891ed5617d59856a6a636c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1343064
Reviewed-by: Steven Grady <slgrady@google.com>
Commit-Queue: Colin Nelson <colnnelson@google.com>
Fuchsia-Auto-Submit: Colin Nelson <colnnelson@google.com>