[build][idk] Ensure sdk_headers_for_internal_use are in headers
Ensure that all files in `sdk_headers_for_internal_use` are also
included in the headers for the library (`public` but sometimes
`sources`). This ensures no header files are missing from the IDK.

Fix violations by removing headers that are not used in the IDK or do
not exist from `sdk_headers_for_internal_use`. Correct the path for one
header in a `zx_library()` instance - "include/" is not implicit as it
is for `sdk_headers`.

Change-Id: Iaeea93a67b6e983b984bfa66fd3146e986f08bb9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1343167
Commit-Queue: David Dorwin <ddorwin@google.com>
Fuchsia-Auto-Submit: David Dorwin <ddorwin@google.com>
Reviewed-by: Jay Zhuang <jayzhuang@google.com>