[kernel][vm] Isolate list is exact
Changes the Isolate list to require that any page marked as being in
the Isolate queue, is in the Isolate list, and similarly any page in
the Isolate list is in the Isolate queue.

Achieving this requires MarkAccessed to handle moving between lists
in this case, instead of just changing queue numbers. As isolate pages
are old and slated for reclamation this is expected to be uncommon, and
measurements show ~1-2% of MarkAccessed calls are against an isolated
page.

As a benefit the Isolate list no longer needs to be processed on LRU
rotation since it cannot have incorrect pages.

Bug: 423900604
Change-Id: Ifed99e80e9b9a6e6b9819ede68ecc8d62c96e4e2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1343244
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>
Reviewed-by: Ethan Williams <ethanws@google.com>
Commit-Queue: Adrian Danis <adanis@google.com>