[kernel][vm] Always Isolate
When making space in the LRU queue always move pages to the Isolate
list instead of to LRU+1. Pages moved into the Isolate list have the
benefit of not needing to be processed next time the LRU queue needs
to be incremented, at the cost of being more expensive to move back
into the general reclaim lists when accessed.

With the additional pages being placed in the Isolate list with this
change the percentage of pages being MarkAccessed that were in the
Isolate state goes from ~1-2% -> ~4%.

Bug: 423900604
Change-Id: I519924e8a8e758674c8ffd1762ffd986caa72a61
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1343245
Reviewed-by: Ethan Williams <ethanws@google.com>
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>
Commit-Queue: Adrian Danis <adanis@google.com>
Fuchsia-Auto-Submit: Adrian Danis <adanis@google.com>