[starnix] Add Linux variant to syscalls generation
The syscall tests will generate two flavors: starnix_ and linux_. The
linux variants use a newly introduced program field, test_target_kernel.
The Starnix Test Runner will later parse this value to determine whether
to run the tests against the typical Starnix environment, or a Machina
guest. The generation of linux_ variants is hidden behind a build arg.

One notable tech debt introduced here is that the Linux-based flavors
do not need to have a Starnix kernel instantiated, and yet the template
will wire up the CML and runner environment to do just that. It's an
unnecessary overhead, but attempting to refactor the syscall generation
further to extract this overhead is beyond what I can commit to...

...pun intended :)

Bug: 414635825
Test: Validate starnix_syscalls_cpp_tests still works.
Change-Id: I685f10b6598274b78cdf06d9f728b001385bcdc8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1335010
Reviewed-by: Guocheng Wei <guochengwei@google.com>
Commit-Queue: Aaron Knobloch <aknobloch@google.com>