[ctf][intl] Removes the flakiness in the `intl` CTF test
This test started failing consistently after a long time of
dormancy. But it turned out that the reason is not mysterious,
though it is a bit unexpected that the flakiness would remain
hidden such a long time.

The reason for flakiness is that there are, in fact, two `OnChange`
requests racing the locale update instead of just one. I added
a wait for the initial `OnChange`, which should detangle the
initialization from the change exercised in the tests.

If we were designing this API again, I suspect we'd want to
remove this opportunity for data race. But since we aren't going
to do this any time soon, treading carefully is another way to
approach this.

In normal operation this API quirk is not that important as we only
really want to observe changes eventually. However, that pattern
makes it hard to write predictable tests without knowing how the
API works under the hood. In this case, the key bit of knowledge
is that `setui` actually sends an initial update to `intl_services`,
a fact that (1) should be inconsequential to begin with (2) means
we need to know the implementation to test the API.

Bug: 383706725
Change-Id: I45c0a43175911bb4eb0bf8ad75bac68ed05a8fd9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1343951
Fuchsia-Auto-Submit: Filip Filmar <fmil@google.com>
Reviewed-by: Christopher Johnson <crjohns@google.com>
Reviewed-by: Brian Bosak <bbosak@google.com>
Commit-Queue: Filip Filmar <fmil@google.com>